Refactor : Network 디렉토리 -> API
