+++ src/Pages/AdminPage/AdminPage.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect, useContext } from 'react';
 
 import { AuthContext } from 'App';
 import { adminCloumns } from 'Utils';
-import { UserInfoService, CRUDUserService } from 'Network';
+import { UserInfoService, CRUDUserService } from 'API';
 
 import { ShowToday, NotValid } from 'Components';
 import SelectedUser from './SelectedUser';
