+++ src/Pages/LoginPage/index.jsx
@@ -2,7 +2,7 @@ import { useState, useContext } from 'react';
 import { useNavigate, Link } from 'react-router-dom';
 
 import { AuthContext } from 'App';
-import { CRUDUserService } from 'Network';
+import { CRUDUserService } from 'API';
 
 import LoginForm from './LoginForm';
 import RegisterForm from './RegisterForm';
