+++ src/Pages/TodoPage/MyTodoList.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { TodoService } from 'Network';
+import { TodoService } from 'API';
 import { format } from 'date-fns';
 import { checkDateTodo } from 'Utils';
